add the invokeTask helper for task composition
