+++ src/dev.task.ts
@@ -1,15 +1,13 @@
 import {Task} from './task/task.js';
-import * as buildTask from './build.task.js';
-import * as serveTask from './serve.task.js';
 
 const DEFAULT_SERVE_DIR = 'dist/';
 
 export const task: Task = {
 	description: 'start development server',
-	run: async (ctx): Promise<void> => {
+	run: async ({args, invokeTask}): Promise<void> => {
 		// TODO fix these
-		ctx.args.watch = true; // TODO always?
-		ctx.args.dir = ctx.args.dir || DEFAULT_SERVE_DIR;
+		args.watch = true; // TODO always?
+		args.dir = args.dir || DEFAULT_SERVE_DIR;
 		// TODO also take HOST and PORT from env
 		// .option('-H, --host', 'Hostname for the server')
 		// .option('-p, --port', 'Port number for the server')
@@ -18,7 +16,7 @@ export const task: Task = {
 		// .option('-w, --watch', 'Watch for changes and rebuild')
 		// .option('-P, --production', 'Set NODE_ENV to production')
 
-		await Promise.all([buildTask.task.run(ctx), serveTask.task.run(ctx)]);
+		await Promise.all([invokeTask('build'), invokeTask('serve')]);
 
 		// ...
 	},
