+++ src/dist.task.ts
@@ -4,16 +4,13 @@ import {paths} from './paths.js';
 import {isTestBuildFile, isTestBuildArtifact} from './oki/testModule.js';
 import {printPath} from './utils/print.js';
 import {cleanDist} from './project/clean.js';
-import {task as assetsTask} from './assets.task.js';
 
 export const isDistFile = (path: string): boolean =>
 	!isTestBuildFile(path) && !isTestBuildArtifact(path);
 
 export const task: Task = {
 	description: 'create the distribution',
-	run: async (ctx) => {
-		const {log} = ctx;
-
+	run: async ({log, invokeTask}) => {
 		await cleanDist(log);
 
 		log.info(`copying ${printPath(paths.build)} to ${printPath(paths.dist)}`);
@@ -21,6 +18,6 @@ export const task: Task = {
 			filter: (id) => isDistFile(id),
 		});
 
-		await assetsTask.run(ctx);
+		await invokeTask('assets');
 	},
 };
