+++ src/project/build.task.ts
@@ -1,6 +1,5 @@
 import {Task} from '../task/task.js';
 import {spawnProcess} from '../utils/process.js';
-import {task as distTask} from './dist.task.js';
 import {cleanBuild} from './clean.js';
 
 // TODO `process.env.NODE_ENV = 'production'`?
@@ -8,14 +7,12 @@ import {cleanBuild} from './clean.js';
 
 export const task: Task = {
 	description: 'build, create, and link the distribution',
-	run: async (ctx) => {
-		const {log} = ctx;
-
+	run: async ({log, invokeTask}) => {
 		await cleanBuild(log);
 
 		log.info('compiling typescript');
 		await spawnProcess('node_modules/.bin/tsc');
 
-		await distTask.run(ctx);
+		await invokeTask('project/dist');
 	},
 };
