+++ src/project/dist.task.ts
@@ -1,19 +1,9 @@
-import {Task, TaskError} from '../task/task.js';
-import {printKeyValue} from '../utils/print.js';
-import {spawnProcess} from '../utils/process.js';
-import {task as distTask} from '../dist.task.js';
+import {Task} from '../task/task.js';
 
 export const task: Task = {
 	description: 'create and link the distribution',
-	run: async (ctx) => {
-		const {log} = ctx;
-
-		await distTask.run(ctx);
-
-		log.info('linking');
-		const linkResult = await spawnProcess('npm', ['link']);
-		if (!linkResult.ok) {
-			throw new TaskError(`Failed to link. ${printKeyValue('code', linkResult.code)}`);
-		}
+	run: async ({invokeTask}) => {
+		await invokeTask('dist');
+		await invokeTask('project/link');
 	},
 };
