+++ src/task/runTask.ts
@@ -1,7 +1,7 @@
 import {SystemLogger} from '../utils/log.js';
-import {cyan, magenta, red} from '../colors/terminal.js';
+import {cyan, magenta, red, gray} from '../colors/terminal.js';
 import {TaskModuleMeta} from './taskModule.js';
-import {Args, Env} from '../cli/types.js';
+import {Args} from '../cli/types.js';
 import {TaskError} from './task.js';
 
 export type RunTaskResult =
@@ -18,14 +18,14 @@ export type RunTaskResult =
 export const runTask = async (
 	task: TaskModuleMeta,
 	args: Args,
-	env: Env,
+	invokeTask: (taskName: string, args: Args) => Promise<void>,
 ): Promise<RunTaskResult> => {
 	let output;
 	try {
 		output = await task.mod.task.run({
 			args,
-			env,
-			log: new SystemLogger([magenta(`[task:${cyan(task.name)}]`)]),
+			log: new SystemLogger([`${gray('[')}${magenta(task.name)}${gray(':log')}${gray(']')}`]),
+			invokeTask: (invokedTaskName, invokedArgs = args) => invokeTask(invokedTaskName, invokedArgs),
 		});
 	} catch (err) {
 		return {
